requirements.txt
sentence-transformers==2.7.0
transformers==4.44.2
torch==2.3.0
faiss-cpu==1.8.0
nltk==3.9b1
pandas==2.2.3
numpy==2.3.4
scikit-learn==1.5.2
tqdm
